requirements.txt
telethon
fastapi
uvicorn
python-dotenv
pydantic
python-multipart